+++ x/btcstaking/keeper/genesis.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"math"
 
-	btcstk "github.com/babylonlabs-io/babylon/btcstaking"
 	bbn "github.com/babylonlabs-io/babylon/types"
 	"github.com/babylonlabs-io/babylon/x/btcstaking/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -27,10 +26,6 @@ func (k Keeper) InitGenesis(ctx context.Context, gs types.GenesisState) error {
 
 	for _, btcDel := range gs.BtcDelegations {
 		k.setBTCDelegation(ctx, btcDel)
-	}
-
-	for _, fpVP := range gs.VotingPowers {
-		k.SetVotingPower(ctx, *fpVP.FpBtcPk, fpVP.BlockHeight, fpVP.VotingPower)
 	}
 
 	for _, blocks := range gs.BlockHeightChains {
@@ -71,11 +66,6 @@ func (k Keeper) ExportGenesis(ctx context.Context) (*types.GenesisState, error) 
 		return nil, err
 	}
 
-	vpFps, err := k.fpVotingPowers(ctx)
-	if err != nil {
-		return nil, err
-	}
-
 	btcDels, err := k.btcDelegators(ctx)
 	if err != nil {
 		return nil, err
@@ -95,7 +85,6 @@ func (k Keeper) ExportGenesis(ctx context.Context) (*types.GenesisState, error) 
 		Params:            k.GetAllParams(ctx),
 		FinalityProviders: fps,
 		BtcDelegations:    dels,
-		VotingPowers:      vpFps,
 		BlockHeightChains: k.blockHeightChains(ctx),
 		BtcDelegators:     btcDels,
 		Events:            evts,
@@ -133,30 +122,6 @@ func (k Keeper) btcDelegations(ctx context.Context) ([]*types.BTCDelegation, err
 	}
 
 	return dels, nil
-}
-
-// fpVotingPowers gets the voting power of a given finality provider at a given Babylon height.
-func (k Keeper) fpVotingPowers(ctx context.Context) ([]*types.VotingPowerFP, error) {
-	iter := k.votingPowerStore(ctx).Iterator(nil, nil)
-	defer iter.Close()
-
-	vpFps := make([]*types.VotingPowerFP, 0)
-
-	for ; iter.Valid(); iter.Next() {
-		blkHeight, fpBTCPK, err := btcstk.ParseBlkHeightAndPubKeyFromStoreKey(iter.Key())
-		if err != nil {
-			return nil, err
-		}
-
-		vp := sdk.BigEndianToUint64(iter.Value())
-		vpFps = append(vpFps, &types.VotingPowerFP{
-			BlockHeight: blkHeight,
-			FpBtcPk:     fpBTCPK,
-			VotingPower: vp,
-		})
-	}
-
-	return vpFps, nil
 }
 
 func (k Keeper) blockHeightChains(ctx context.Context) []*types.BlockHeightBbnToBtc {
