+++ x/finality/keeper/genesis.go
@@ -48,6 +48,10 @@ func (k Keeper) InitGenesis(ctx context.Context, gs types.GenesisState) error {
 		}
 	}
 
+	for _, fpVP := range gs.VotingPowers {
+		k.SetVotingPower(ctx, *fpVP.FpBtcPk, fpVP.BlockHeight, fpVP.VotingPower)
+	}
+
 	return k.SetParams(ctx, gs.Params)
 }
 
@@ -83,6 +87,11 @@ func (k Keeper) ExportGenesis(ctx context.Context) (*types.GenesisState, error) 
 		return nil, err
 	}
 
+	vpFps, err := k.fpVotingPowers(ctx)
+	if err != nil {
+		return nil, err
+	}
+
 	return &types.GenesisState{
 		Params:           k.GetParams(ctx),
 		IndexedBlocks:    blocks,
@@ -92,6 +101,7 @@ func (k Keeper) ExportGenesis(ctx context.Context) (*types.GenesisState, error) 
 		PubRandCommit:    prCommit,
 		SigningInfos:     signingInfos,
 		MissedBlocks:     missedBlocks,
+		VotingPowers:     vpFps,
 	}, nil
 }
 
@@ -252,6 +262,30 @@ func (k Keeper) signingInfosAndMissedBlock(ctx context.Context) ([]types.Signing
 	}
 
 	return signingInfos, missedBlocks, nil
+}
+
+// fpVotingPowers gets the voting power of a given finality provider at a given Babylon height.
+func (k Keeper) fpVotingPowers(ctx context.Context) ([]*types.VotingPowerFP, error) {
+	iter := k.votingPowerStore(ctx).Iterator(nil, nil)
+	defer iter.Close()
+
+	vpFps := make([]*types.VotingPowerFP, 0)
+
+	for ; iter.Valid(); iter.Next() {
+		blkHeight, fpBTCPK, err := btcstk.ParseBlkHeightAndPubKeyFromStoreKey(iter.Key())
+		if err != nil {
+			return nil, err
+		}
+
+		vp := sdk.BigEndianToUint64(iter.Value())
+		vpFps = append(vpFps, &types.VotingPowerFP{
+			BlockHeight: blkHeight,
+			FpBtcPk:     fpBTCPK,
+			VotingPower: vp,
+		})
+	}
+
+	return vpFps, nil
 }
 
 // parsePubKeyAndBlkHeightFromStoreKey expects to receive a key with
